1. Moving components to their folders
