+++ interview-scheduler/src/index.js
@@ -1,12 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './components/main/App';
 import { Provider } from 'react-redux';
 import store from './store/store';
-
-
-
 
 ReactDOM.render(
     <Provider store={store}>
